Fix test

+++ MediaBrowser.Providers/Manager/ItemImageProvider.cs
@@ -381,7 +381,7 @@ namespace MediaBrowser.Providers.Manager
                     {
                         // Always remove empty parent folder
                         var folder = Path.GetDirectoryName(image.Path);
-                        if (!_fileSystem.GetFiles(folder).Any())
+                        if (Directory.Exists(folder) && !_fileSystem.GetFiles(folder).Any())
                         {
                             Directory.Delete(folder);
                         }
